- Added some tickcount things (I hope this works, as I tested many things regarding skills :( )  - Added pet showup in the cashshop window  - Changed and checking the portal count, just like MS  - Might fixed something else, idk, I'm sleepy...
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@2981 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/CashServer/CashServer.cpp
@@ -62,7 +62,7 @@ void CashServer::loadLogConfig() {
 }
 
 string CashServer::makeLogIdentifier() {
-	return "World: " + boost::lexical_cast<string>(getWorld()) + "; Cash server";
+	return "World: " + boost::lexical_cast<string>(static_cast<int16_t>(getWorld())) + "; Cash server";
 }
 
 void CashServer::connectWorld() {
